feature(agent,api,pkg): add VPN capability
It adds to the ShellHub's Agent the capability to connect to a
ShellHub's Enterprise service, which provides a virtual private network
between devices registered into the same namespace.

To enable it, the ShellHub's instance must support it, and the
`SHELLHUB_VPN` environmental variable must be set to `TRUE` on the
ShellHub Agent startup.